move state files up
