+++ react-sdk/src/providers/tambo-prop-stream-provider/loading.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { useTamboStream } from "../provider";
-import { DEFAULT_STREAM_KEY, StreamStateComponentProps } from "../types";
+import { useTamboStream } from "./provider";
+import { DEFAULT_STREAM_KEY, StreamStateComponentProps } from "./types";
 
 /**
  * Loading component that renders children when the stream is in a loading state
